Initial refactoring efforts to remove costly file scanning
