+++ src/Providers/BaseModuleServiceProvider.php
@@ -1,0 +1,18 @@
+<?php
+namespace LangleyFoxall\Modules\Providers;
+
+use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
+
+class BaseModuleServiceProvider extends ServiceProvider
+{
+    /** @var string[] $providers */
+    protected $providers = [];
+
+    public function register()
+    {
+        foreach($this->providers as $provider) {
+            $this->app->register($provider);
+        }
+    }
+
+}
