+++ src/Template/Providers/ModuleServiceProvider.php
@@ -1,0 +1,12 @@
+<?php
+namespace LangleyFoxall\Modules\Template\Providers;
+
+use LangleyFoxall\Modules\Providers\BaseModuleServiceProvider;
+
+class ModuleServiceProvider extends BaseModuleServiceProvider
+{
+    /** @var string[] $providers */
+    protected $providers = [
+        RouteServiceProvider::class
+    ];
+}
