+++ src/Template/Providers/RouteServiceProvider.php
@@ -1,5 +1,5 @@
 <?php
-namespace LangleyFoxall\Modules\Template;
+namespace LangleyFoxall\Modules\Template\Providers;
 
 use Illuminate\Support\Facades\Route;
 use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
